Update unit tests and re organise files
